Rename types -> typings
Having a file named the same as a standard python module cause a bug in
my dev setup.
